Revert r110506: undo change to phrasing on Special:UploadWizard's most important button
The big button label was changed from 'Select a media file to donate' to 'Drop media file to donate here' to make it "clear that it's possible to drag and drop multiple files".
Unfortunately this is probably premature; the drag-n-drop isn't very slickly integrated yet and for instance has no visual indication of whether you're dropping in the right stop. If you get it wrong -- very easy -- then you replace the entire page with just your image in the browser window (bug 34667).
Switching it back to the original form.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk/extensions@112517 dd0e9695-b195-4be7-bd10-2dea1a65a6b6